requirements.txt
# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# Speech-to-text
deepgram-sdk>=3.0.0,<4.0.0

# Web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0

